Cleaned up Sprint 2 a little. I moved the commands that were in Game 1 to the keyboard controller and added comments throughout the item files.
