+++ Keyboard/KeyboardCont.cs
@@ -22,14 +22,17 @@ namespace LegendOfZelda
         // constructor
         public KeyboardCont(Game1 game)
         {
+            // Controller mappings holds the on press transition commands
             controllerMappings = new Dictionary<Keys, ICommand>();
+            // these two are more self explanatory maps for while keys are held down and when a key is released
             heldDownMappings = new Dictionary<Keys, ICommand>();
             releaseMappings = new Dictionary<Keys, ICommand>();
 
             myGame = game;
+            //keeps track of keys that were already pressed down to detect transition.
             alrPressed = new List<Keys>();
             
-            // set up the table
+            // set up the table which can be later moved to a content loader
             ICommand c = new CommQuit(game);
             ICommand a = new CommLinkMove(game, new Vector2(-1, 0));
             ICommand d = new CommLinkMove(game, new Vector2(1, 0));
@@ -56,6 +59,11 @@ namespace LegendOfZelda
             RegisterCommand(Keys.D2 , boomerang);
             RegisterCommand(Keys.I, nextItem);
             RegisterCommand(Keys.U, lastItem);
+
+            ICommand pSC = new PreviousSpriteCommand(game);
+            ICommand nSC = new NextSpriteCommand(game);
+            RegisterCommand(Keys.O, pSC);
+            RegisterCommand(Keys.P, nSC);
 
 
             // set up held down table
@@ -92,19 +100,23 @@ namespace LegendOfZelda
 
         public void Update()
         {
+            // get the keys that are currently pressed
             Keys[] pressedKeys = Keyboard.GetState().GetPressedKeys();
             foreach (Keys key in pressedKeys)
             {
-
+                // if the key has an on transition command mapped to it
                 if (controllerMappings.ContainsKey(key))
                 {
+                    //and its transitioning and not being held
                     if (!alrPressed.Contains(key))
                     {
+                        //execute the on transition command and add it to alr Pressed so that the transition command wont re-trigger
                         controllerMappings[key].Execute();
                         alrPressed.Add(key);
                     }
 
                 }
+                //also any key that is being pressed should execute its held down command if it has one
                 if (heldDownMappings.ContainsKey(key))
                 {
                     heldDownMappings[key].Execute();
@@ -112,6 +124,7 @@ namespace LegendOfZelda
             }
             // remove unpressed keys
 
+            //get a list of keys that were released
             unPressList = new List<Keys>();
             foreach (Keys key in alrPressed)
             {
@@ -120,6 +133,7 @@ namespace LegendOfZelda
                     unPressList.Add(key);
                 }
             }
+            // execute their commands if they had them and remove them from the list of keys that were pressed
             foreach (Keys key in unPressList)
             {
                 // This executes the command on key release
